+++ src/Swarm/Game/Scenario.hs
@@ -69,6 +69,7 @@ import Data.Set qualified as Set
 import Data.Text (Text)
 import Data.Text qualified as T
 import Swarm.Game.Entity
+import Swarm.Game.Entity.Specimens (worldAttributes)
 import Swarm.Game.Failure
 import Swarm.Game.Location
 import Swarm.Game.Recipe
@@ -90,7 +91,6 @@ import Swarm.Language.Pipeline (ProcessedTerm)
 import Swarm.Language.Pretty (prettyText)
 import Swarm.Language.Syntax (Syntax)
 import Swarm.Language.Text.Markdown (Document)
-import Swarm.TUI.View.Attribute.Attr (worldAttributeNames)
 import Swarm.TUI.View.Attribute.CustomStyling (toAttrPair)
 import Swarm.Util (binTuples, failT)
 import Swarm.Util.Effect (ignoreWarnings, throwToMaybe, withThrow)
@@ -150,7 +150,7 @@ instance FromJSONE (EntityMap, WorldMap) Scenario where
     emRaw <- liftE (v .:? "entities" .!= [])
 
     parsedAttrs <- liftE (v .:? "attrs" .!= [])
-    let attrsUnion = Set.fromList (map (fst . toAttrPair) parsedAttrs) <> worldAttributeNames
+    let attrsUnion = Set.fromList (map (fst . toAttrPair) parsedAttrs) <> M.keysSet worldAttributes
 
     em <- case run . runThrow $ buildEntityMap attrsUnion emRaw of
       Right x -> return x
