+++ swarm.cabal
@@ -118,6 +118,7 @@ library
                       Swarm.Game.Display
                       Swarm.Game.Entity
                       Swarm.Game.Entity.Cosmetic
+                      Swarm.Game.Entity.Specimens
                       Swarm.Game.Exception
                       Swarm.Game.Location
                       Swarm.Game.Recipe
